// packages/sdk-typescript/src/generated/types/participantType.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from "@solana/kit";

/** Participant type enum for distinguishing between humans and agents */
export enum ParticipantType {
  Human,
  Agent,
}

export type ParticipantTypeArgs = ParticipantType;

export function getParticipantTypeEncoder(): FixedSizeEncoder<ParticipantTypeArgs> {
  return getEnumEncoder(ParticipantType);
}

export function getParticipantTypeDecoder(): FixedSizeDecoder<ParticipantType> {
  return getEnumDecoder(ParticipantType);
}

export function getParticipantTypeCodec(): FixedSizeCodec<
  ParticipantTypeArgs,
  ParticipantType
> {
  return combineCodec(getParticipantTypeEncoder(), getParticipantTypeDecoder());
}